Moved all configuration classes from CLI folder to Configuration.
