+++ deploy/src/Microsoft.Azure.IIoT.Deployment/Configuration/ConsoleConfigurationProvider.cs
@@ -8,7 +8,6 @@ namespace Microsoft.Azure.IIoT.Deployment.Configuration {
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using Microsoft.Azure.IIoT.Deployment.Cli;
     using Microsoft.Azure.Management.ResourceManager.Fluent;
     using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
