+++ deploy/src/Microsoft.Azure.IIoT.Deployment/Microsoft.Azure.IIoT.Deployment.csproj
@@ -44,8 +44,6 @@
     <PackageReference Include="KubernetesClient" Version="1.6.4" />
 
     <PackageReference Include="System.Private.Uri" Version="4.3.2" />
-
-    <PackageReference Include="CommandLineParser" Version="2.6.0" />
   </ItemGroup>
 
   <ItemGroup>
